requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pandas==2.2.3
matplotlib==3.8.1
seaborn==0.13.2